schedule attribute to person added as a single compulsory field
