+++ src/seedu/addressbook/commands/ClearCommand.java
@@ -14,8 +14,7 @@ public class ClearCommand extends Command {
     @Override
     public CommandResult execute() {
         addressBook.clear();
-        commandHistory.checkForAction();
-        commandHistory.addHistory(COMMAND_WORD);
+        addressBook.checkForAction();
         return new CommandResult(MESSAGE_SUCCESS);
     }
 }
