+++ src/seedu/addressbook/commands/Command.java
@@ -2,7 +2,6 @@ package seedu.addressbook.commands;
 
 import seedu.addressbook.common.Messages;
 import seedu.addressbook.data.AddressBook;
-import seedu.addressbook.data.CommandHistory;
 import seedu.addressbook.data.person.ReadOnlyPerson;
 
 import java.util.List;
@@ -14,7 +13,6 @@ import static seedu.addressbook.ui.Gui.DISPLAYED_INDEX_OFFSET;
  */
 public abstract class Command {
     protected AddressBook addressBook;
-    protected CommandHistory commandHistory;
     protected List<? extends ReadOnlyPerson> relevantPersons;
     private int targetIndex = -1;
 
@@ -53,7 +51,6 @@ public abstract class Command {
      */
     public void setData(AddressBook addressBook, List<? extends ReadOnlyPerson> relevantPersons) {
         this.addressBook = addressBook;
-        this.commandHistory = addressBook.getCommandHistory();
         this.relevantPersons = relevantPersons;
     }
