+++ src/seedu/addressbook/commands/DeleteCommand.java
@@ -30,8 +30,7 @@ public class DeleteCommand extends Command {
         try {
             final ReadOnlyPerson target = getTargetPerson();
             addressBook.removePerson(target);
-            commandHistory.checkForAction();
-            commandHistory.addHistory(COMMAND_WORD + " " + getTargetIndex());
+            addressBook.checkForAction();
             return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, target));
 
         } catch (IndexOutOfBoundsException ie) {
