+++ src/seedu/addressbook/commands/HelpCommand.java
@@ -20,8 +20,7 @@ public class HelpCommand extends Command {
             + "\n" + ViewAllCommand.MESSAGE_USAGE
             + "\n" + HelpCommand.MESSAGE_USAGE
             + "\n" + ExitCommand.MESSAGE_USAGE
-            + "\n" + UndoCommand.MESSAGE_USAGE
-            + "\n" + RedoCommand.MESSAGE_USAGE;
+            + "\n" + UndoCommand.MESSAGE_USAGE;
 
     @Override
     public CommandResult execute() {
