+++ src/seedu/addressbook/commands/ViewAllCommand.java
@@ -32,7 +32,6 @@ public class ViewAllCommand extends Command {
             if (!addressBook.containsPerson(target)) {
                 return new CommandResult(Messages.MESSAGE_PERSON_NOT_IN_ADDRESSBOOK);
             }
-            commandHistory.addHistory(COMMAND_WORD + " " + getTargetIndex());
             return new CommandResult(String.format(MESSAGE_VIEW_PERSON_DETAILS, target.getAsTextShowAll()));
         } catch (IndexOutOfBoundsException ie) {
             return new CommandResult(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
