+++ src/seedu/addressbook/data/AddressBook.java
@@ -6,13 +6,17 @@ import seedu.addressbook.data.person.UniquePersonList;
 import seedu.addressbook.data.person.UniquePersonList.DuplicatePersonException;
 import seedu.addressbook.data.person.UniquePersonList.PersonNotFoundException;
 
+import java.util.ArrayList;
+
 /**
  * Represents the entire address book. Contains the data of the address book.
  */
 public class AddressBook {
 
-    private CommandHistory commandHistory;
     private UniquePersonList allPersons;
+    private ArrayList<UniquePersonList> addressBookStates = new ArrayList<>();
+    private int STATEITERATOR = 0;
+    private int ITERATOROFFSET = 1;
 
     public static AddressBook empty() {
         return new AddressBook();
@@ -23,7 +27,6 @@ public class AddressBook {
      */
     public AddressBook() {
         allPersons = new UniquePersonList();
-        commandHistory = new CommandHistory(this);
     }
 
     /**
@@ -33,7 +36,7 @@ public class AddressBook {
      */
     public AddressBook(UniquePersonList persons) {
         this.allPersons = new UniquePersonList(persons);
-        this.commandHistory = new CommandHistory(this);
+        saveInitialState();
     }
 
     public void switchAddressBook(UniquePersonList persons){
@@ -91,7 +94,65 @@ public class AddressBook {
         return allPersons.hashCode();
     }
 
-    public CommandHistory getCommandHistory() {
-        return commandHistory;
+    public static class HistoryOutOfBoundException extends Exception {}
+    /**
+     * Save initial state of the address book
+     */
+    public void saveInitialState() {
+        this.addressBookStates.add(getAllPersons());
+    }
+
+    /**
+     * add the current state of the address book into the history
+     */
+    public void addCurrentState() {
+        moveIteratorForward();
+        addressBookStates.add(getAllPersons());
+    }
+
+    public void moveIteratorForward() {
+        STATEITERATOR++;
+    }
+
+    public void moveIteratorBackward() {
+        STATEITERATOR--;
+    }
+
+    /**
+     * truncate old history when new undo-able command is executed after undo-ing
+     */
+    public void truncateOldPath() {
+        addressBookStates.subList(STATEITERATOR + ITERATOROFFSET, addressBookStates.size()).clear();
+    }
+
+    /**
+     * check whether to execute truncateOldPath
+     */
+    public void checkForAction() {
+        if(!nextBoxIsEmpty()) truncateOldPath();
+        addCurrentState();
+    }
+
+    public boolean nextBoxIsEmpty() {
+        if((STATEITERATOR + ITERATOROFFSET) >= addressBookStates.size()) return true;
+        else return false;
+    }
+
+    public void undoLast() throws HistoryOutOfBoundException {
+        moveIteratorBackward();
+        if(STATEITERATOR < 0){
+            moveIteratorForward();
+            throw new HistoryOutOfBoundException();
+        }
+        switchAddressBook(addressBookStates.get(STATEITERATOR));
+    }
+
+    public void redoLast() throws HistoryOutOfBoundException {
+        moveIteratorForward();
+        if(STATEITERATOR + ITERATOROFFSET> addressBookStates.size()){
+            moveIteratorBackward();
+            throw new HistoryOutOfBoundException();
+        }
+        switchAddressBook(addressBookStates.get(STATEITERATOR));
     }
 }
