+++ src/seedu/addressbook/data/person/Person.java
@@ -16,20 +16,16 @@ public class Person implements ReadOnlyPerson {
     private Phone phone;
     private Email email;
     private Address address;
-    private Title title;
-    private Schedule schedule;
 
     private final Set<Tag> tags = new HashSet<>();
     /**
      * Assumption: Every field must be present and not null.
      */
-    public Person(Name name, Phone phone, Email email, Address address, Title title, Schedule schedule, Set<Tag> tags) {
+    public Person(Name name, Phone phone, Email email, Address address, Set<Tag> tags) {
         this.name = name;
         this.phone = phone;
         this.email = email;
         this.address = address;
-        this.title = title;
-        this.schedule = schedule;
         this.tags.addAll(tags);
     }
 
@@ -37,7 +33,7 @@ public class Person implements ReadOnlyPerson {
      * Copy constructor.
      */
     public Person(ReadOnlyPerson source) {
-        this(source.getName(), source.getPhone(), source.getEmail(), source.getAddress(), source.getTitle(), source.getSchedule(), source.getTags());
+        this(source.getName(), source.getPhone(), source.getEmail(), source.getAddress(), source.getTags());
     }
 
     @Override
@@ -59,12 +55,6 @@ public class Person implements ReadOnlyPerson {
     public Address getAddress() {
         return address;
     }
-
-    @Override
-    public Title getTitle() { return title; }
-
-    @Override
-    public Schedule getSchedule() { return schedule; }
 
     @Override
     public Set<Tag> getTags() {
@@ -89,7 +79,7 @@ public class Person implements ReadOnlyPerson {
     @Override
     public int hashCode() {
         // use this method for custom fields hashing instead of implementing your own
-        return Objects.hash(name, phone, email, address, title, schedule, tags);
+        return Objects.hash(name, phone, email, address, tags);
     }
 
     @Override
