+++ src/seedu/addressbook/data/person/ReadOnlyPerson.java
@@ -14,8 +14,6 @@ public interface ReadOnlyPerson {
     Phone getPhone();
     Email getEmail();
     Address getAddress();
-    Title getTitle();
-    Schedule getSchedule();
 
     /**
      * The returned {@code Set} is a deep copy of the internal {@code Set},
@@ -32,9 +30,7 @@ public interface ReadOnlyPerson {
                 && other.getName().equals(this.getName()) // state checks here onwards
                 && other.getPhone().equals(this.getPhone())
                 && other.getEmail().equals(this.getEmail())
-                && other.getAddress().equals(this.getAddress())
-                && other.getTitle().equals(this.getTitle())
-                && other.getSchedule().equals(this.getSchedule()));
+                && other.getAddress().equals(this.getAddress()));
     }
 
     /**
@@ -59,16 +55,6 @@ public interface ReadOnlyPerson {
             builder.append(detailIsPrivate);
         }
         builder.append(getAddress())
-                .append(" Title: ");
-        if (getTitle().isPrivate()) {
-            builder.append(detailIsPrivate);
-        }
-        builder.append(getTitle())
-                .append(" Schedule: ");
-        if (getSchedule().isPrivate()) {
-            builder.append(detailIsPrivate);
-        }
-        builder.append(getSchedule())
                 .append(" Tags: ");
         for (Tag tag : getTags()) {
             builder.append(tag);
@@ -90,12 +76,6 @@ public interface ReadOnlyPerson {
         }
         if (!getAddress().isPrivate()) {
             builder.append(" Address: ").append(getAddress());
-        }
-        if (!getTitle().isPrivate()) {
-            builder.append(" Title: ").append(getTitle());
-        }
-        if (!getSchedule().isPrivate()) {
-            builder.append(" Schedule: ").append(getSchedule());
         }
         builder.append(" Tags: ");
         for (Tag tag : getTags()) {
