+++ src/seedu/addressbook/storage/jaxb/AdaptedPerson.java
@@ -33,11 +33,6 @@ public class AdaptedPerson {
     private AdaptedContactDetail email;
     @XmlElement(required = true)
     private AdaptedContactDetail address;
-    @XmlElement(required = true)
-    private AdaptedContactDetail title;
-
-    @XmlElement(required = true)
-    private AdaptedContactDetail schedule;
 
     @XmlElement
     private List<AdaptedTag> tagged = new ArrayList<>();
@@ -68,14 +63,6 @@ public class AdaptedPerson {
         address.isPrivate = source.getAddress().isPrivate();
         address.value = source.getAddress().value;
 
-        title = new AdaptedContactDetail();
-        title.isPrivate = source.getTitle().isPrivate();
-        title.value = source.getTitle().value;
-
-        schedule = new AdaptedContactDetail();
-        schedule.isPrivate = source.getSchedule().isPrivate();
-        schedule.value = source.getSchedule().value;
-
         tagged = new ArrayList<>();
         for (Tag tag : source.getTags()) {
             tagged.add(new AdaptedTag(tag));
@@ -97,8 +84,8 @@ public class AdaptedPerson {
             }
         }
         // second call only happens if phone/email/address are all not null
-        return Utils.isAnyNull(name, phone, email, address, title, schedule)
-                || Utils.isAnyNull(phone.value, email.value, address.value, title.value, schedule.value);
+        return Utils.isAnyNull(name, phone, email, address)
+                || Utils.isAnyNull(phone.value, email.value, address.value);
     }
 
     /**
@@ -115,8 +102,6 @@ public class AdaptedPerson {
         final Phone phone = new Phone(this.phone.value, this.phone.isPrivate);
         final Email email = new Email(this.email.value, this.email.isPrivate);
         final Address address = new Address(this.address.value, this.address.isPrivate);
-        final Title title = new Title(this.title.value, this.title.isPrivate);
-        final Schedule schedule = new Schedule(this.schedule.value, this.schedule.isPrivate);
-        return new Person(name, phone, email, address, title, schedule, tags);
+        return new Person(name, phone, email, address, tags);
     }
 }
