+++ test/java/seedu/addressbook/parser/ParserTest.java
@@ -252,8 +252,6 @@ public class ParserTest {
                 new Phone(Phone.EXAMPLE, true),
                 new Email(Email.EXAMPLE, false),
                 new Address(Address.EXAMPLE, true),
-                new Title(Title.EXAMPLE, true),
-                new Schedule(Schedule.EXAMPLE, true),
                 new HashSet<>(Arrays.asList(new Tag("tag1"), new Tag("tag2"), new Tag("tag3")))
             );
         } catch (IllegalValueException ive) {
@@ -266,9 +264,7 @@ public class ParserTest {
                 + person.getName().fullName
                 + (person.getPhone().isPrivate() ? " pp/" : " p/") + person.getPhone().value
                 + (person.getEmail().isPrivate() ? " pe/" : " e/") + person.getEmail().value
-                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().value
-                + (person.getTitle().isPrivate() ? " ps/" : " s/") + person.getTitle().value
-                + (person.getSchedule().isPrivate() ? " pd/" : " d/") + person.getSchedule().value;
+                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().value;
         for (Tag tag : person.getTags()) {
             addCommand += " t/" + tag.tagName;
         }
